Consolidate type checks to single file
