+++ src/NodeProxy.js
@@ -4,7 +4,7 @@ import document from 'global/document'
 import {descriptors} from './propertyDescriptors'
 import {addEventListener, removeEventListener} from './eventDelegator'
 import {emitMount, emitUnmount} from './mountable'
-import {isDefined} from './isDefined'
+import {isDefined} from './is'
 import {get} from './get'
 import {set} from './set'
