--- src/isDefined.js
@@ -1,5 +1,0 @@
-/* @flow */
-
-export function isDefined (obj?: any): boolean {
-  return (typeof obj !== `undefined`)
-}
