--- src/isEmptyObject.js
@@ -1,5 +1,0 @@
-/* @flow */
-
-export function isEmptyObject (obj: Object): boolean {
-  return Object.keys(obj).length === 0
-}
