--- src/isFunction.js
@@ -1,5 +1,0 @@
-/* @flow */
-
-export function isFunction (fn?: any): boolean {
-  return Object.prototype.toString.call(fn) === `[object Function]`
-}
