--- src/isNumber.js
@@ -1,5 +1,0 @@
-/* @flow */
-
-export function isNumber (obj?: any): boolean {
-  return (typeof obj === `number`)
-}
