--- src/isObservable.js
@@ -1,7 +1,0 @@
-/* @flow */
-
-import {Observable} from 'rxjs/Observable'
-
-export function isObservable (obj: any): boolean {
-  return obj && obj instanceof Observable
-}
