--- src/isString.js
@@ -1,5 +1,0 @@
-/* @flow */
-
-export function isString (obj?: any): boolean {
-  return (typeof obj === `string`)
-}
