--- src/isSubject.js
@@ -1,7 +1,0 @@
-/* @flow */
-
-import {Subject} from 'rxjs/Subject'
-
-export function isSubject (obj: any): boolean {
-  return Boolean(obj instanceof Subject)
-}
