--- src/isVirtual.js
@@ -1,7 +1,0 @@
-/* @flow */
-
-import {$$virtual} from './symbol'
-
-export function isVirtual (obj?: any): boolean {
-  return !!obj && obj[$$virtual]
-}
