+++ src/mountable.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import {CustomEvent} from './CustomEvent'
-import {isFunction} from './isFunction'
+import {isFunction} from './is'
 
 export function emitMount (node: HTMLElement, fn: Function | void): void {
   if (isFunction(fn) && node.parentNode) {
